"Finishing" touches
Made it so going to .test page goes straight to log in, and home page now has links to navigate the site (browse movies, browse by category). Also just in case added middleware to any routes that lead to changing database entries. Now technically all functionality is present in project. I'll probably keep messing around to refactor/get further acquainted with the framework, but for the purpose of this challenge, it is complete!

+++ movie-collection/routes/web.php
@@ -11,31 +11,34 @@
 |
 */
 
-Route::get('/', function () {
-    return view('welcome');
-});
+// Route::get('/', function () {
+//     return view('welcome');
+// });
 
 Auth::routes();
 
+Route::get('/', 'HomeController@index')->name('home');
+
 Route::get('/home', 'HomeController@index')->name('home');
 
-//Movie Routes
-//Used to show list of movies
+// Movie Routes
+// ->name assigns name to route, ->middleware makes user sign in to access route
+// Used to show list of movies
 Route::get('/movies', 'MovieController@index')->name('movies.index');
 
-//Used to create movies
-Route::get('/movies/create', 'MovieController@create');
+// Used to create movies
+Route::get('/movies/create', 'MovieController@create')->middleware('auth');
 Route::post('/movies', 'MovieController@store');
 
-//Used to show specific movie info
+// Used to show specific movie info
 Route::get('/movies/{movie}', 'MovieController@show')->name('movies.show');
 
-//Used to edit/update movie
-Route::get('/movies/{movie}/edit', 'MovieController@edit');
+// Used to edit/update movie
+Route::get('/movies/{movie}/edit', 'MovieController@edit')->middleware('auth');
 Route::put('/movies/{movie}', 'MovieController@update');
 
-//Called by a form on movie.show page, deletes current movie
-Route::delete('movies/{movie}', 'MovieController@destroy')->name('movies.destroy');
+// Called by a form on movie.show page, deletes current movie
+Route::delete('movies/{movie}', 'MovieController@destroy')->name('movies.destroy')->middleware('auth');
 
-//Category Route
+// Category Route
 Route::get('/categories', 'CategoryController@index');
